requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
PyJWT==2.6.0
werkzeug==2.2.2
